Add function(s) to calculate if point intersects rhombus
Created function and helper functions and classes to calculate if a
given point intersects a rhombus. This was so that the hover to shared
mvoes behaviour could be implemented - we have to manually check which
lines actually intersect the mousepoint as the built in method returns
all lines who's containing box intersect.